Refactoring helper functions into utils folder
